Adding Test for ConcreteRegularPolygon

+++ Polygons.Library/ConcreteRegularPolygon.cs
@@ -3,9 +3,9 @@
     public class ConcreteRegularPolygon
     {
         public int NumberOfSides { get; set; }  
-        public int SideLength { get; set; }
+        public double SideLength { get; set; }
 
-        public ConcreteRegularPolygon(int sides, int length)
+        public ConcreteRegularPolygon(int sides, double length)
         {
             NumberOfSides = sides;
             SideLength = length;
